DRLND/navigation/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.10.0=pypi_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=pypi_0
atomicwrites=1.4.0=pypi_0
attrs=20.2.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bleach=3.1.5=pypi_0
cattrs=1.0.0=pypi_0
certifi=2020.6.20=py36_0
cffi=1.14.2=pypi_0
cloudpickle=1.3.0=pypi_0
colorama=0.4.3=pypi_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.3=pypi_0
freetype=2.10.2=hd328e21_0
future=0.18.2=pypi_0
gast=0.4.0=pypi_0
grpcio=1.11.0=pypi_0
gym=0.17.2=pypi_0
h5py=2.10.0=pypi_0
html5lib=0.9999999=pypi_0
importlib-metadata=1.7.0=pypi_0
iniconfig=1.0.1=pypi_0
intel-openmp=2020.2=254
ipykernel=5.3.4=pypi_0
ipython=7.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-console=6.2.0=pypi_0
jupyter-core=4.6.3=pypi_0
kiwisolver=1.2.0=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=h62dcd97_1
markdown=3.2.2=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36hb782905_0
mkl_fft=1.1.0=py36h45dec08_0
mkl_random=1.1.1=py36h47e9c7a_0
mlagents=0.19.0=pypi_0
mlagents-envs=0.19.0=pypi_0
more-itertools=8.5.0=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.7=pypi_0
ninja=1.10.1=py36h7ef1ec2_0
notebook=6.1.3=pypi_0
numpy=1.19.1=py36h5510c5b_0
numpy-base=1.19.1=py36ha3acd2a_0
olefile=0.46=py_0
packaging=20.4=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.7.1=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.2.0=py36hcc1f983_0
pip=20.2.2=py36_0
pluggy=0.13.1=pypi_0
prometheus-client=0.8.0=pypi_0
prompt-toolkit=3.0.7=pypi_0
protobuf=3.5.2=pypi_0
py=1.9.0=pypi_0
pycparser=2.20=pypi_0
pyglet=1.5.0=pypi_0
pygments=2.6.1=pypi_0
pyparsing=2.4.7=pypi_0
pypiwin32=223=pypi_0
pyrsistent=0.16.0=pypi_0
pytest=6.0.1=pypi_0
python=3.6.10=h9f7ef89_2
python-dateutil=2.8.1=pypi_0
pytorch=1.6.0=py3.6_cuda102_cudnn7_0
pywin32=228=pypi_0
pywinpty=0.5.7=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=pypi_0
qtconsole=4.7.7=pypi_0
qtpy=1.9.0=pypi_0
scipy=1.5.2=pypi_0
send2trash=1.5.0=pypi_0
setuptools=49.6.0=py36_0
six=1.15.0=py_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=1.7.0=pypi_0
tensorflow=1.7.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
tk=8.6.10=he774522_0
toml=0.10.1=pypi_0
torchvision=0.7.0=py36_cu102
tornado=6.0.4=pypi_0
traitlets=4.3.3=pypi_0
unityagents=0.4.0=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=pypi_0
wincertstore=0.2=py36h7fe50ca_0
xz=5.2.5=h62dcd97_0
zipp=3.1.0=pypi_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
